Rename default package to `client.ts`
And drop `test` subdirectory as it doesn't help much.
