+++ package.json
@@ -2,18 +2,18 @@
   "name": "smee-client",
   "version": "1.2.4",
   "description": "Client to proxy webhooks to local host",
-  "main": "index.js",
+  "main": "client.ts",
   "bin": {
     "smee": "./bin/smee.js"
   },
   "files": [
-    "index.js",
-    "index.d.ts",
+    "client.js",
+    "client.d.ts",
     "bin"
   ],
   "scripts": {
     "test": "jest --coverage && standard",
-    "lint": "eslint index.ts bin/smee.js test/index.test.ts",
+    "lint": "eslint client.ts bin/smee.js client.test.ts",
     "build": "tsc -p tsconfig.json"
   },
   "repository": "github:distributed-system-analysis/smee-client",
