+++ tsconfig.json
@@ -59,5 +59,5 @@
     "skipLibCheck": true
   },
   "include": ["./*"],
-  "files": ["index.ts", "./test/index.test.ts"]
+  "files": ["client.ts", "client.test.ts"]
 }
